incomplete refactor, wip
