--- examples/function/udfproj/entry.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -eux
-
-python3 example.py
