--- examples/sink/simplesink/entry.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -eux
-
-python example.py
