+++ pynumaflow/function/__init__.py
@@ -1,4 +1,7 @@
 from pynumaflow.function._dtypes import Message, Messages
 from pynumaflow.function.handler import HTTPHandler
+from pynumaflow.function.server import UserDefinedFunctionServicer
+from pynumaflow.function import udfunction_pb2
+import udfunction_pb2
 
-__all__ = ["Message", "Messages", "HTTPHandler"]
+# __all__ = ["Message", "Messages", "HTTPHandler", "UserDefinedFunctionServicer"]
